asdasd

+++ src/main/java/br/com/app/gym/web/service/CheckinService.java
@@ -2,6 +2,7 @@ package br.com.app.gym.web.service;
 
 import br.com.app.gym.web.model.CheckinDadosCliente;
 import br.com.app.gym.web.model.CheckinSolicitacao;
+import java.math.BigDecimal;
 import java.util.List;
 import javax.ws.rs.ClientErrorException;
 
@@ -14,6 +15,6 @@ public interface CheckinService {
 
     public CheckinDadosCliente getDadosCliente(Integer checkinId, Integer academiaId) throws ClientErrorException;
 
-    public boolean liberarCliente(Integer checkinId, Integer clienteId, Integer academiaId, boolean liberado, Integer servicoId) throws ClientErrorException;
+    public boolean liberarCliente(Integer checkinId, Integer clienteId, Integer academiaId, boolean liberado, BigDecimal servicoId) throws ClientErrorException;
 
 }
